Move HashableByteArray to decaton.processor.internal
